feat : price range slider 이동 시 search bar 상태 변경 및 초기화 버튼 구현
